requirements.txt
1password
openai
requests
python-dotenv
nest_asyncio